+++ Examples/SmokeTest/SmokeTest/ErrorHelpers/CatchNSException.h
@@ -1,0 +1,9 @@
+//
+//  Use this file to import your target's public headers that you would like to expose to Swift.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface CatchNSException: NSObject
++ (NSException *)throwAndCatchNSException;
+@end