+++ Examples/SmokeTest/SmokeTest/ErrorHelpers/CatchNSException.m
@@ -1,0 +1,18 @@
+#import "CatchNSException.h"
+
+@implementation CatchNSException
+
++ (NSException *)throwAndCatchNSException {
+    NSException * exception = nil;
+    @try {
+        @throw [NSException exceptionWithName:@"TestException" reason:@"Exception Handling reason" userInfo: nil];
+    }
+    @catch(NSException * caughtException) {
+        exception = caughtException;
+    }
+    @finally {
+        return exception;
+    }
+}
+
+@end